requirements.txt
spotipy
pynput
python-dotenv
SpeechRecognition
pyaudio
traceback
tkinter